UPDATE: date_diff test updated to Long return type
git-svn-id: https://svn.javlin.eu/svn/cloveretl.opensource/trunk@8776 a09ad3ba-1a0f-0410-b1b9-c67202f10d70

+++ cloveretl.engine/test/org/jetel/ctl/CompilerTestCase.java
@@ -2117,7 +2117,7 @@ public abstract class CompilerTestCase extends CloverTestCase {
 		doCompile("test_datelib_dateDiff");
 		Calendar cal = Calendar.getInstance();
 		cal.setTime(BORN_VALUE);
-		int diff = cal.get(Calendar.YEAR);
+		long diff = cal.get(Calendar.YEAR);
 		cal.setTime(new Date());
 		diff -= cal.get(Calendar.YEAR);
 		check("ddiff", diff);
